fix: add user profile endpoints to admin user profile controller
