+++ API/CCW.UserProfile/Controllers/AdminUserController.cs
@@ -87,6 +87,69 @@ public class AdminUserController : ControllerBase
         }
     }
 
+    [Authorize(Policy = "B2CUsers")]
+    [Route("createUserProfile")]
+    [HttpPut]
+    public async Task<IActionResult> CreateUserProfile([FromBody] UserProfileRequestModel request)
+    {
+        try
+        {
+            GetUserId(out var userId);
+            User newUser = _mapper.Map<User>(request);
+            newUser.Id = userId;
+            var userEmail = User.Claims.FirstOrDefault(c => c.Type == "emails")?.Value;
+            newUser.Email = userEmail;
+            var createdUser = await _cosmosDbService.AddUserAsync(newUser, cancellationToken: default);
+
+            return Ok(_mapper.Map<UserProfileResponseModel>(createdUser));
+        }
+        catch (Exception e)
+        {
+            var originalException = e.GetBaseException();
+            _logger.LogError(originalException, originalException.Message);
+            return NotFound("An error occur while trying to create new user.");
+        }
+    }
+
+    [Authorize(Policy = "B2CUsers")]
+    [Route("getUserProfile")]
+    [HttpGet]
+    public async Task<IActionResult> GetUserProfile()
+    {
+        try
+        {
+            GetUserId(out var userId);
+            var result = await _cosmosDbService.GetUserAsync(userId, cancellationToken: default);
+
+            return (result != null) ? Ok(_mapper.Map<UserProfileResponseModel>(result)) : NotFound();
+        }
+        catch (Exception e)
+        {
+            var originalException = e.GetBaseException();
+            _logger.LogError(originalException, originalException.Message);
+            return NotFound("An error occur while trying to retrieve user.");
+        }
+    }
+
+    [Authorize(Policy = "AADUsers")]
+    [Route("getUnmatchedUserProfiles")]
+    [HttpGet]
+    public async Task<IActionResult> GetUnmatchedUserProfiles()
+    {
+        try
+        {
+            List<User> result = await _cosmosDbService.GetUnmatchedUserProfiles(cancellationToken: default);
+
+            return Ok(result);
+        }
+        catch (Exception e)
+        {
+            var originalException = e.GetBaseException();
+            _logger.LogError(originalException, originalException.Message);
+            return NotFound("An error occur while trying to get all unmatched user profiles");
+        }
+    }
+
     private void GetUserId(out string userId)
     {
         userId = this.HttpContext.User.Claims
